test(cli): consolidate socket cli tests into test_client.py for 1:1 structure conformance
